refactor(container): mv GlobalLayout to layout scope
