+++ src/containers/layout/GlobalLayout/dynamic.js
