+++ src/containers/layout/GlobalLayout/tests/store.test.js
