+++ src/pages/_error.js
@@ -9,7 +9,7 @@ import ErrorPage from '@/components/ErrorPage'
 
 import { useStore } from '@/stores/init'
 
-import { Wrapper } from '@/containers/GlobalLayout/styles'
+import { Wrapper } from '@/containers/layout/GlobalLayout/styles'
 
 const Error = (props) => {
   const store = useStore(props)
