+++ src/pages/meetups.js
@@ -8,7 +8,7 @@ import { ROUTE } from '@/constant'
 import { getJwtToken, makeGQClient, ssrAmbulance, parseTheme } from '@/utils'
 import { P } from '@/schemas'
 
-import GlobalLayout from '@/containers/GlobalLayout'
+import GlobalLayout from '@/containers/layout/GlobalLayout'
 import MeetupsContent from '@/containers/content/MeetupsContent'
 
 import { useStore } from '@/stores/init'
