+++ src/pages/sponsor.js
@@ -7,7 +7,7 @@ import { ROUTE } from '@/constant'
 
 import { getJwtToken, makeGQClient, ssrAmbulance, parseTheme } from '@/utils'
 import { P } from '@/schemas'
-import GlobalLayout from '@/containers/GlobalLayout'
+import GlobalLayout from '@/containers/layout/GlobalLayout'
 import SponsorContent from '@/containers/content/SponsorContent'
 
 import { useStore } from '@/stores/init'
